+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 * Implement details view for mapping profiles (UIDATIMP-219)
 * Implement create mapping profile feature (UIDATIMP-227)
 * Implement edit mapping profile feature (UIDATIMP-230)
+* Implement mapping profile duplication feature (UIDATIMP-228)
+* Implement mapping profile deletion feature (UIDATIMP-229)
 
 ## [1.3.0](https://github.com/folio-org/ui-data-import/tree/v1.3.0) (2019-06-12)
 * Create data import settings page's 3rd pane for Match Profiles. Add string capitalization (capitalize) and HTML special chars decoding (htmlDecode) utils (UIDATIMP-6)